refactor(subnets): barrel exports
