+++ src/app/subnets/views/Subnets/components/AddFabric/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddFabric";
