+++ src/app/subnets/views/Subnets/components/AddSpace/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddSpace";
