+++ src/app/subnets/views/Subnets/components/AddSubnet/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddSubnet";
