+++ src/app/subnets/views/Subnets/components/AddVlan/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddVlan";
