+++ src/app/subnets/views/Subnets/components/StaticDHCPLease/StaticDHCPTable/useStaticDHCPTableColumns/useStaticDHCPTableColumns.tsx
@@ -3,7 +3,7 @@ import { useMemo } from "react";
 import type { ColumnDef } from "@tanstack/react-table";
 import { Link } from "react-router";
 
-import { SubnetDetailsSidePanelViews } from "../../../../views/SubnetDetails/constants";
+import { SubnetDetailsSidePanelViews } from "../../../../views/constants";
 
 import TableActions from "@/app/base/components/TableActions";
 import { useSidePanel } from "@/app/base/side-panel-context";
