+++ src/app/subnets/views/Subnets/components/SubnetActionForms/components/DeleteSubnet/DeleteSubnet.tsx
@@ -9,7 +9,7 @@ import { subnetActions } from "@/app/store/subnet";
 import { useCanBeDeleted, useIsDHCPEnabled } from "@/app/store/subnet/hooks";
 import subnetSelectors from "@/app/store/subnet/selectors";
 import subnetURLs from "@/app/subnets/urls";
-import type { SubnetActionProps } from "@/app/subnets/views/Subnets/views/SubnetDetails/types";
+import type { SubnetActionProps } from "@/app/subnets/views/Subnets/views/types";
 
 export const DeleteSubnet = ({
   subnetId,
