+++ src/app/subnets/views/Subnets/components/SubnetDetailsHeader/SubnetDetailsHeader.tsx
@@ -9,7 +9,7 @@ import {
   subnetActionLabels,
   SubnetActionTypes,
   SubnetDetailsSidePanelViews,
-} from "@/app/subnets/views/Subnets/views/SubnetDetails/constants";
+} from "@/app/subnets/views/Subnets/views/constants";
 
 type Props = {
   subnet: Subnet;
