+++ src/app/subnets/views/Subnets/components/SubnetsListHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SubnetsListHeader";
