+++ src/app/subnets/views/Subnets/views/SubnetsList/SubnetsList.tsx
@@ -1,12 +1,13 @@
 import type { ReactElement } from "react";
 import { useState } from "react";
 
-import SubnetsTable from "../../components/SubnetsTable";
-
 import PageContent from "@/app/base/components/PageContent/PageContent";
 import { useWindowTitle } from "@/app/base/hooks";
 import { useQuery } from "@/app/base/hooks/urls";
-import SubnetsListHeader from "@/app/subnets/views/Subnets/components/SubnetsListHeader/SubnetsListHeader";
+import {
+  SubnetsListHeader,
+  SubnetsTable,
+} from "@/app/subnets/views/Subnets/components";
 import type { GroupByKey } from "@/app/subnets/views/Subnets/components/SubnetsTable/types";
 
 export const SubnetsUrlParams = {
