+++ src/app/subnets/views/Subnets/views/constants.ts
