+++ src/app/subnets/views/Subnets/views/types.ts
